refactor: correct typo in employee model import
